Refactor MacAppSessionController into extensions
